[Data management] Added Network detection class for repositories. Added Room to the app, with one table for customers (declaration will change in the future). Added Dao class for customers room table. Added Dao inheritance for common database operations.
Added local and remote repository for customers, and a defaultRepository to choose between them.
Added an interface for customers repositories, to be able to mock these repositories when testing the app.

Added Results wrapper class for repositories.

[Refactor]
Changed ViewModel instance declaration, and ViewModelFactory implementation. Added ViewModel injection implementation to Injection object.

[UI]
Created an adapter to display customers list in CustomerFragment RecyclerView. Updated CustomerFragment to configure ViewModel and RecyclerView.
Updated CustomerFragment to be able to communicate with parent activity, when user click on the Floating Action Button.
Changes with HomeFragment, this is the second button that launch CustomerFragment

[Tests]
Added DefaultCustomerRepository test class.
Added database Dao tests for Customers table.
Added utility class for LiveData tests.
TargetSDKVersion is now 28, to be able to use Robolectric tests (targeting android 29 need to use Java 9, and the project and CI/CD aren't ready for Java 9.