[View] Added base fragment class for data binding. Added UI components to insert a new customer.
[Model]
Added observe method in repositories, to be able to observe customers list changes in database.
Added methods to save a customer or a list of customers in database.
Updated repository to get informations from local repository, for now.
Database changes (missing suspend modifiers in DAO interface).

[Tests]
Updated tests to pass with changed code in repositories.
Removed activityTest, it will be re worked at the end of development